move input-output toggle (and checkbox) into repl package
