+++ packages/repl/src/lib/Output/index.svelte
@@ -119,7 +119,7 @@
 
 <style>
 	.view-toggle {
-		height: var(--pane-controls-h);
+		height: 4.2rem;
 		border-bottom: 1px solid #eee;
 		white-space: nowrap;
 		box-sizing: border-box;
