+++ packages/repl/src/lib/index.svelte
@@ -2,6 +2,7 @@
 	import { setContext, createEventDispatcher } from 'svelte';
 	import { writable } from 'svelte/store';
 	import SplitPane from './SplitPane.svelte';
+	import InputOutputToggle from './InputOutputToggle.svelte';
 	import ComponentSelector from './Input/ComponentSelector.svelte';
 	import ModuleEditor from './Input/ModuleEditor.svelte';
 	import Output from './Output/index.svelte';
@@ -97,6 +98,9 @@
 
 	let module_editor;
 	let output;
+
+	let width = 0;
+	let show_output = false;
 
 	let current_token;
 	async function rebundle() {
@@ -240,25 +244,33 @@
 	$: if (output && $selected) {
 		output.update($selected, $compile_options);
 	}
+
+	$: mobile = width < 540;
 </script>
 
 <svelte:window on:beforeunload={beforeUnload} />
 
-<div class="container" class:orientation>
-	<SplitPane
-		type={orientation === 'rows' ? 'vertical' : 'horizontal'}
-		pos={fixed ? fixedPos : orientation === 'rows' ? 50 : 60}
-		{fixed}
-	>
-		<section slot="a">
-			<ComponentSelector show_modified={showModified} {handle_select} on:add on:remove />
-			<ModuleEditor errorLoc={sourceErrorLoc} {theme} />
-		</section>
+<div class="container" class:mobile bind:clientWidth={width}>
+	<div class="viewport" class:output={show_output}>
+		<SplitPane
+			type={orientation === 'rows' ? 'vertical' : 'horizontal'}
+			pos={mobile ? fixedPos : orientation === 'rows' ? 50 : 60}
+			fixed={mobile}
+		>
+			<section slot="a">
+				<ComponentSelector show_modified={showModified} {handle_select} on:add on:remove />
+				<ModuleEditor errorLoc={sourceErrorLoc} {theme} />
+			</section>
 
-		<section slot="b" style="height: 100%;">
-			<Output {svelteUrl} status={status_visible && status} {embedded} {relaxed} {injectedJS} {injectedCSS} {theme} />
-		</section>
-	</SplitPane>
+			<section slot="b" style="height: 100%;">
+				<Output {svelteUrl} status={status_visible && status} {embedded} {relaxed} {injectedJS} {injectedCSS} {theme} />
+			</section>
+		</SplitPane>
+	</div>
+
+	{#if mobile}
+		<InputOutputToggle bind:checked={show_output}/>
+	{/if}
 </div>
 
 <style>
@@ -287,5 +299,19 @@
 	.container :global(section) > :global(*):last-child {
 		width: 100%;
 		height: 100%;
+	}
+
+	.viewport {
+		height: 100%;
+	}
+
+	.mobile .viewport {
+		width: 200%;
+		height: calc(100% - 42px);
+		transition: transform 0.3s;
+	}
+
+	.mobile .viewport.output {
+		transform: translate(-50%);
 	}
 </style>
